+++ FirstCRUDApplication/Repositories/IDbRepository.cs
@@ -6,7 +6,7 @@ namespace Coffee.Repositories
     public interface IDbRepository<TEntity> where TEntity : class
     {
         void Create(TEntity item);
-        TEntity FindById(int id);
+        TEntity FindById(long id);
         IEnumerable<TEntity> Get();
         IEnumerable<TEntity> Get(Func<TEntity, bool> predicate);
         void Remove(TEntity item);
